+++ src/main/java/com/talha/academix/services/impl/ExamServiceImpl.java
@@ -8,10 +8,10 @@ import org.springframework.stereotype.Service;
 import com.talha.academix.dto.ExamDTO;
 import com.talha.academix.exception.ResourceNotFoundException;
 import com.talha.academix.exception.RoleMismatchException;
-import com.talha.academix.model.Course;
 import com.talha.academix.model.Exam;
 import com.talha.academix.repository.CourseRepo;
 import com.talha.academix.repository.ExamRepo;
+import com.talha.academix.services.CourseService;
 import com.talha.academix.services.ExamService;
 
 import lombok.RequiredArgsConstructor;
@@ -22,34 +22,37 @@ public class ExamServiceImpl implements ExamService {
 
     private final ExamRepo examRepo;
     private final CourseRepo courseRepo;
+    private final CourseService courseService;
     private final ModelMapper modelMapper;
 
     @Override
     public ExamDTO createExam(Long teacherId, ExamDTO dto) {
 
-       Course course = courseRepo.findById(dto.getCourseId())
-       .orElseThrow(()-> new ResourceNotFoundException("Course not found with id: " + dto.getCourseId()));
+        if (courseService.teacherValidation(teacherId, dto.getCourseId())) {
 
-       if(!course.getTeacher().equals(teacherId)){
-           throw new RoleMismatchException("Unauthorized Teacher");
-       }
+            Exam exam = modelMapper.map(dto, Exam.class);
+            exam.setCourse(courseRepo.findById(dto.getCourseId())
+            .orElseThrow(() -> new ResourceNotFoundException("Course not found")));        
+            exam = examRepo.save(exam);
 
-       Exam exam = modelMapper.map(dto, Exam.class);
-       exam.setCourse(course);
-       exam = examRepo.save(exam);
-
-       return modelMapper.map(exam, ExamDTO.class);
+            return modelMapper.map(exam, ExamDTO.class);
+        } else
+            throw new RoleMismatchException("only Teacher can create exam");
     }
 
     @Override
-    public ExamDTO updateExam(Long examId, ExamDTO dto) {
+    public ExamDTO updateExam(Long userid, Long examId, ExamDTO dto) {
         Exam existing = examRepo.findById(examId)
-            .orElseThrow(() -> new ResourceNotFoundException("Exam not found with id: " + examId));
+                .orElseThrow(() -> new ResourceNotFoundException("Exam not found with id: " + examId));
 
-        existing.setTitle(dto.getTitle());
-        examRepo.save(existing);
+        if (courseService.teacherValidation(userid, existing.getCourse().getCourseid())) {
+            existing.setTitle(dto.getTitle());
+            examRepo.save(existing);
 
-        return modelMapper.map(existing, ExamDTO.class);
+            return modelMapper.map(existing, ExamDTO.class);
+        } else {
+            throw new RoleMismatchException("Only teacher can update exam");
+        }
     }
 
     @Override
@@ -63,14 +66,19 @@ public class ExamServiceImpl implements ExamService {
     public List<ExamDTO> getExamsByCourse(Long courseId) {
         List<Exam> exams = examRepo.findByCourseId(courseId);
         return exams.stream()
-        .map(e -> modelMapper.map(e, ExamDTO.class))
-        .toList();
+                .map(e -> modelMapper.map(e, ExamDTO.class))
+                .toList();
     }
 
     @Override
-    public void deleteExam(Long examId) {
+    public void deleteExam(Long userid, Long examId) {
         Exam exam = examRepo.findById(examId)
                 .orElseThrow(() -> new ResourceNotFoundException("Exam not found with id: " + examId));
-        examRepo.delete(exam);
-    }    
+
+        if (courseService.teacherValidation(userid, exam.getCourse().getCourseid())) {
+            examRepo.delete(exam);
+        } else {
+            throw new RoleMismatchException("Only teacher can delete exam");
+        }
+    }
 }
