+++ src/main/java/com/talha/academix/services/impl/LectureServiceImpl.java
@@ -8,10 +8,12 @@ import org.springframework.stereotype.Service;
 
 import com.talha.academix.dto.LectureDTO;
 import com.talha.academix.exception.ResourceNotFoundException;
+import com.talha.academix.exception.RoleMismatchException;
 import com.talha.academix.model.Content;
 import com.talha.academix.model.Lecture;
 import com.talha.academix.repository.ContentRepo;
 import com.talha.academix.repository.LectureRepo;
+import com.talha.academix.services.CourseService;
 import com.talha.academix.services.LectureService;
 
 import lombok.RequiredArgsConstructor;
@@ -22,40 +24,55 @@ public class LectureServiceImpl implements LectureService {
 
     private final LectureRepo lectureRepo;
     private final ContentRepo contentRepo;
+    private final CourseService courseService;
     private final ModelMapper mapper;
 
     @Override
     public LectureDTO addLecture(Long userid, LectureDTO dto) {
+    
         Content content = contentRepo.findById(dto.getContentId())
-            .orElseThrow(() -> new ResourceNotFoundException("Content not found: " + dto.getContentId()));
+        .orElseThrow(() -> new ResourceNotFoundException("Content not found: " + dto.getContentId()));
 
-        Lecture lecture = new Lecture();
-        lecture.setContent(content);
-        lecture.setTitle(dto.getTitle());
-        lecture.setVideoUrl(dto.getVideoUrl());
-        lecture.setDuration(dto.getDuration());
-        lecture = lectureRepo.save(lecture);
+        if(courseService.teacherValidation(userid, content.getCourse().getCourseid())){
 
-        return mapper.map(lecture, LectureDTO.class);
+            Lecture lecture = new Lecture();
+            lecture.setContent(content);
+            lecture.setTitle(dto.getTitle());
+            lecture.setVideoUrl(dto.getVideoUrl());
+            lecture.setDuration(dto.getDuration());
+            lecture = lectureRepo.save(lecture);
+    
+            return mapper.map(lecture, LectureDTO.class);
+        }
+        else throw new RoleMismatchException("only valid user can add leature");
     }
 
     @Override
     public LectureDTO updateLecture(Long userid, Long lectureId, LectureDTO dto) {
-        Lecture existing = lectureRepo.findById(lectureId)
-            .orElseThrow(() -> new ResourceNotFoundException("Lecture not found: " + lectureId));
 
+        Content content = contentRepo.findById(dto.getContentId())
+        .orElseThrow(() -> new ResourceNotFoundException("Content not found: " + dto.getContentId()));
+
+        if(courseService.teacherValidation(userid, content.getCourse().getCourseid())){
+
+        Lecture existing = lectureRepo.findById(lectureId)
+        .orElseThrow(() -> new ResourceNotFoundException("Lecture not found: " + lectureId));
+
+        
         // if contentId changed, reassign content
-        if (!existing.getContent().getContentID().equals(dto.getContentId())) {
-            Content content = contentRepo.findById(dto.getContentId())
+        if (!existing.getContent().getContentID().equals(content.getContentID())) {
+            Content newcontent = contentRepo.findById(dto.getContentId())
                 .orElseThrow(() -> new ResourceNotFoundException("Content not found: " + dto.getContentId()));
-            existing.setContent(content);
+            existing.setContent(newcontent);
         }
+
         existing.setTitle(dto.getTitle());
         existing.setVideoUrl(dto.getVideoUrl());
         existing.setDuration(dto.getDuration());
         existing = lectureRepo.save(existing);
 
-        return mapper.map(existing, LectureDTO.class);
+        return mapper.map(existing, LectureDTO.class);}
+        else throw new RoleMismatchException("only teacher can update lectures");
     }
 
     @Override
@@ -76,8 +93,16 @@ public class LectureServiceImpl implements LectureService {
 
     @Override
     public void deleteLecture(Long userid, Long lectureId) {
+    
         Lecture lecture = lectureRepo.findById(lectureId)
             .orElseThrow(() -> new ResourceNotFoundException("Lecture not found: " + lectureId));
+
+        Content content = contentRepo.findById(lecture.getContent().getContentID())
+            .orElseThrow(() -> new ResourceNotFoundException("COntent not found"));
+
+        if(courseService.teacherValidation(userid, content.getCourse().getCourseid())){
         lectureRepo.delete(lecture);
+        }
+        else throw new RoleMismatchException("Only Teacher can delete lecture");
     }
 }
